Fixed Error Handling for integrateImage energy
When attempting to integrate data from cycle 2022-2, was getting an
IndexError when integrateImageStack called integrateSingleImage.  The
indexes for finding the energy were switched from what they needed to
be. Perhaps this block of code isn't running for other cycles, or the
index structure is different.

Changes are at lines 62 and 63.